requirements.txt
annotated-types==0.7.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.0
